[CASCL-1323] Consolidate `install_mode.go` in `apply` like the other pflag types
The earlier split of `install_mode.go` into a bare type in `apply/`
and a pflag wrapper in `install/` added boilerplate (a wrapper type,
a conversion at every call site) without buying clarity. Move the
whole file to `apply/` like `create_karpenter_resources.go` and
`inference_method.go` already are: the pflag methods live where the
type lives, and the `install` cobra command binds `apply.InstallMode`
directly.

`update` does not bind `--install-mode`, so the pflag methods just go
unused there — same situation as `apply.InstallMode` itself before
this commit. No behavior change.
