+++ cmd/kubectl-datadog/autoscaling/cluster/apply/install_mode.go
@@ -1,6 +1,8 @@
 package apply
 
 import (
+	"fmt"
+
 	"github.com/DataDog/datadog-operator/cmd/kubectl-datadog/autoscaling/cluster/common/aws"
 )
 
@@ -18,6 +20,30 @@ const (
 // install-mode. Stacks created before this tag was introduced have no tag and
 // are treated as install-mode=existing-nodes.
 const InstallModeTagKey = "install-mode"
+
+// String returns the string representation of the InstallMode.
+func (i *InstallMode) String() string {
+	return string(*i)
+}
+
+// Set sets the InstallMode value from a string.
+func (i *InstallMode) Set(s string) error {
+	switch s {
+	case "fargate":
+		*i = InstallModeFargate
+	case "existing-nodes":
+		*i = InstallModeExistingNodes
+	default:
+		return fmt.Errorf("install-mode must be one of fargate or existing-nodes")
+	}
+
+	return nil
+}
+
+// Type returns the type name for pflag.
+func (*InstallMode) Type() string {
+	return "InstallMode"
+}
 
 // DetectedInstallMode reads the install-mode tag from a CFN stack. Stacks
 // created before this tag was introduced have no tag and default to
